statistical modules moved to new sub-package 'stats'
