+++ qats/signal.py
