+++ qats/stats/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+"""
+Sub-package for statistics/distributions.
+"""
