+++ qats/stats/weibull.py
@@ -6,8 +6,8 @@ import numpy as np
 from scipy.special import gamma, binom
 from scipy.optimize import leastsq, fsolve, brentq
 import matplotlib.pyplot as plt
-from .stats import empirical_cdf
-from .signal import find_maxima
+from .empirical import empirical_cdf
+from ..signal import find_maxima
 
 
 # todo: build documentation and check that docstrings behave as intended
